code/lidar/requirements.txt
PyYAML
numpy
opencv-python
mayavi
open3d